qcom-cpufreq: Notify the current frequency to the topology driver
The topology driver keeps track of the current capacity of a CPU
and make it available via arch_scale_freq_capacity(). The PELT
load tracking in the scheduler and energy-aware task placement
are the clients of this API. Hence notify the new frequency to the
topology driver during frequency transition.

Change-Id: I6dae674ba3b9d4c745b3caff59b165215926d970
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Anush02198 <Anush.4376@gmail.com>